backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lat_lab"
version = "0.1.0"
description = "LAT-LAB 博客平台"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "LAT-LAB团队", email = "example@example.com"},
]
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.10.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "aiofiles>=23.1.0",
    "Jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "pymysql>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "flake8>=6.0.0",
    "pre-commit>=3.3.1",
]

[project.scripts]
lat-lab = "src.lat_lab.main:run_app"

[tool.hatch.build.targets.wheel]
packages = ["src/lat_lab"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"